+++ test/browser/index.html
@@ -12,6 +12,15 @@
   <div id="mocha"></div>
   <script src="../../bower_components/mocha/mocha.js"></script>
 
+  <!--[if lt IE 10]>
+    <script type="text/javascript" src="../../plugin/web-socket-js/swfobject.js"></script>
+    <script type="text/javascript" src="../../plugin/web-socket-js/web_socket.js"></script>
+    <script type="text/javascript">
+    // 设置变量，配置插件中 WebSocketMain.swf 的引用路径
+    WEB_SOCKET_SWF_LOCATION = "../../plugin/web-socket-js/WebSocketMain.swf";
+    </script>
+  <![endif]-->
+
   <script>
     mocha.setup({
       ui: 'bdd',
